Move rate limiter creation into root helper
